Refactor EC2MacLauncher/EC2UnixLauncher/EC2WindowsSSHLauncher to EC2SSHLauncher
